requirements.txt
pyAesCrypt==6.0.0
python-dotenv==1.0.0
setuptools==67.3.2
